+++ be15-jamjam-bookjeok-msa/be15-2nd-jamjam-bookjeok/src/main/java/com/jamjam/bookjeok/auth/service/AuthService.java
@@ -5,7 +5,7 @@ import com.jamjam.bookjeok.auth.dto.TokenResponse;
 import com.jamjam.bookjeok.auth.entity.RefreshToken;
 import com.jamjam.bookjeok.auth.repository.RefreshTokenRepository;
 import com.jamjam.bookjeok.domains.member.command.repository.MemberRepository;
-import com.jamjam.bookjeok.domains.member.entity.Member;
+import com.jamjam.bookjeok.domains.member.command.entity.Member;
 
 import com.jamjam.bookjeok.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
